pyproject.toml
[project]
name = "deepseek-from-scratch"
version = "0.1.0"
description = "DeepSeek-V3 implementation from scratch with Ray pipeline orchestration"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
authors = [
    { name = "DeepSeek Team" }
]
keywords = ["deepseek", "ray", "training", "mlx", "pytorch", "rust", "llm", "transformer"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]

dependencies = [
    # Core
    "numpy>=1.24.0",
    "pyyaml>=6.0",
    "tqdm>=4.65.0",
    "rich>=13.0.0",
    
    # Ray framework
    "ray[train,data]>=2.9.0",
    
    # CLI
    "typer>=0.9.0",
    
    # PyTorch & Transformers
    "torch>=2.0.0",
    "transformers>=4.35.0",
    "tokenizers>=0.15.0",
    "datasets>=2.14.0",
    "accelerate>=0.24.0",
    "safetensors>=0.4.0",
    
    # Logging & Monitoring
    "structlog>=25.5.0",
    "prometheus-client>=0.23.1",
    
    # Environment management
    "python-dotenv>=1.0.0",
    
    # Modal cloud GPU
    "modal>=0.64.0",
    
    # MLX (Apple Silicon) - optional at runtime
    "mlx>=0.5.0; sys_platform == 'darwin' and platform_machine == 'arm64'",
]

[project.optional-dependencies]
pytorch = [
    "torch>=2.0.0",
    "transformers>=4.35.0",
    "tokenizers>=0.15.0",
    "datasets>=2.14.0",
    "accelerate>=0.24.0",
]
mlx = [
    "mlx>=0.5.0",
]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "pytest-asyncio>=0.21.0",
    "mypy>=1.5.0",
    "ruff>=0.1.0",
    "pre-commit>=3.4.0",
]
modal = [
    "modal>=0.64.0",
    "torch>=2.0.0",
    "transformers>=4.35.0",
    "tokenizers>=0.15.0",
    "datasets>=2.14.0",
    "accelerate>=0.24.0",
    "gguf>=0.6.0",
]
all = [
    "deepseek-from-scratch[pytorch,mlx,modal,dev]",
]

[project.scripts]
deepseek-pipeline = "ray_pipeline.cli:app"

[project.urls]
Homepage = "https://github.com/deepseek-ai/deepseek-from-scratch"
Repository = "https://github.com/deepseek-ai/deepseek-from-scratch"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["ray_pipeline", "deepseek-from-scratch-python/src/deepseek", "deepseek-from-scratch-python/mlx_impl"]

[tool.uv]
package = true
managed = true
dev-dependencies = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "mypy>=1.5.0",
    "ruff>=0.1.0",
]

[tool.uv.sources]
# Use uv for all dependency management
# Install: uv sync
# Install with Modal: uv sync --extra modal
# Install all extras: uv sync --all-extras

[tool.ruff]
target-version = "py310"
line-length = 100
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # Pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade
]
ignore = [
    "E501",   # line too long
    "B008",   # do not perform function calls in argument defaults
    "B905",   # zip without strict
]

[tool.ruff.per-file-ignores]
"__init__.py" = ["F401"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
exclude = [
    "tests/",
    "build/",
    "dist/",
    "Deepseek-from-scratch-in-rust/",
]

[[tool.mypy.overrides]]
module = [
    "ray.*",
    "torch.*",
    "mlx.*",
    "transformers.*",
    "datasets.*",
]
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["deepseek-from-scratch-python/tests", "ray_pipeline/tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short"

[tool.coverage.run]
source = ["ray_pipeline", "deepseek-from-scratch-python/src"]
omit = ["*/tests/*", "*/__pycache__/*"]
